split out racing TestMonitorHasSignerParameters
